# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Hyperledger Explorer REST API Swagger",
    "description": "Rest API for fabric .",
    "version": "1.0.0",
    "contact": {
      "name": "Hyperledger Team",
      "email": ""
    }
  },
  "host": "localhost:8080",
  "schemes": ["http"],
  "basePath": "/api",
  "produces": ["application/json"],
...
  "paths": {
    "/channels": {
      "get": {
        "tags": ["Channel"],
        "summary": "Query to fetch channels",
        "description": "Query to fetch channels",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Channels"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/curChannel": {
      "get": {
        "tags": ["Channel"],
        "summary": "get Current channel",
        "description": "get Current channel",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/curChannel"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/changeChannel/{channel_genesis_hash}": {
      "get": {
        "tags": ["Channel"],
        "summary": "Query to get Change Channel on a channel",
        "description": "Query to get Change Channel on a channel",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/changeChannel"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/chaincode/{channel_genesis_hash}": {
      "get": {
        "tags": ["Chaincodes"],
        "summary": "Query to fetch all Installed/instantiated chaincodes",
        "description": "Query to fetch all Installed/instantiated chaincodes",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "version": {
                    "type": "string"
                  },
                  "path": {
                    "type": "string"
                  },
                  "channel_genesis_hash": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/channel": {
      "post": {
        "tags": ["Channel"],
        "summary": "Create New channel on target peers  ",
        "description": "Create New channel on target peers",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "channelName",
            "in": "formData",
            "description": "channel name",
            "required": true,
            "type": "string"
          },
          {
            "name": "genesisBlock",
            "in": "formData",
            "description": "gensis block",
            "required": true,
            "type": "string"
          },
          {
            "name": "orgName",
            "in": "formData",
            "description": "organation name ",
            "required": true,
            "type": "string"
          },
          {
            "name": "channelArtificats",
            "in": "formData",
            "description": "channelArtificats",
            "required": true,
            "type": "array",
            "items": {
              "type": "file"
            }
          },
          {
            "name": "profile",
            "in": "formData",
            "description": "block profile",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/joinChannel": {
      "post": {
        "tags": ["Channel"],
        "summary": "join channel ",
        "description": "join  channel ",
        "consumes": ["application/json", "application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "payload",
            "schema": {
              "type": "object",
              "required": true,
              "properties": {
                "channelName": {
                  "type": "string"
                },
                "orgName": {
                  "type": "string"
                },
                "peers": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/block/{channel_genesis_hash}/{number}": {
      "get": {
        "tags": ["Blocks"],
        "summary": "Query to Fetch Blocks",
        "description": "Query to Fetch Blocks",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          },
          {
            "name": "number",
            "in": "path",
            "description": "block number",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "number": {
                  "type": "string"
                },
                "previous_hash": {
                  "type": "string"
                },
                "channel_genesis_hash": {
                  "type": "string"
                },
                "data_hash": {
                  "type": "string"
                },
                "transactions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "signature": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string"
                          },
                          "data": {
                            "type": "array",
                            "items": {
                              "type": "number"
                            }
                          }
                        }
                      },
                      "payload": {
                        "$ref": "#/definitions/blockPayload"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/peersStatus/{channelName}": {
      "get": {
        "tags": ["Channel"],
        "summary": "Query Get PeersStatus",
        "description": "Query Get PeersStatus",
        "parameters": [
          {
            "name": "channelName",
            "in": "path",
            "description": "channel name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "peers": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "requests": {
                        "type": "string"
                      },
                      "server_hostname": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Channels": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "type": "array",
        "items": {
          "type": "integer"
        }
      }
    },
    "curChannel": {
      "type": "object",
      "properties": {
        "currentChannel": {
          "type": "string"
        }
      }
    },
    "blockPayload": {
      "type": "object",
      "properties": {
        "header": {
          "type": "object",
          "properties": {
            "channel_header": {
              "type": "object",
              "properties": {
                "version": {
                  "type": "number"
                },
                "type": {
                  "type": "number"
                },
                "time_stamp": {
                  "type": "string"
                },
                "channel_id": {
                  "type": "string"
                },
                "txid": {
                  "type": "string"
                },
                "epoch": {
                  "type": "string"
                }
              }
            }
          },
          "signature_header": {
            "type": "object",
            "properties": {
              "creator": {
                "type": "object",
                "properties": {
                  "Mspid": {
                    "type": "string"
                  },
                  "Idbytes": {
                    "type": "string"
                  }
                }
              },
              "nonce": {
                "type": "object",
                "properties": {
                  "data": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    }
                  },
                  "type": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "data": {
          "type": "object"
        }
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}